requirements.txt
pocket-api
wallabag_api
